+++ packages/xrpl/HISTORY.md
@@ -5,8 +5,13 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 ## Unreleased
 * Remove references to the Hooks testnet faucet in the xrpl.js code repository.
 
+### BREAKING CHANGES
+* Use rippled api_version v2 as default while maintaining support for v1.
+
 ### Added
 * Add `nfts_by_issuer` clio-only API definition
+* Support for the `fixPreviousTxnID` amendment.
+
 ## 3.1.0 (2024-06-03)
 
 ### BREAKING CHANGES